Move kafka notifications into a module
Will be adding another notification module and want to have them in the
same parent module
