+++ test_manual/kafka_notifier_manual_test.py
@@ -23,7 +23,7 @@ from datetime import datetime, timezone
 import sys
 import time
 
-from cdmtaskservice.kafka_notifications import KafkaNotifier
+from cdmtaskservice.notifications.kafka_notifications import KafkaNotifier
 from cdmtaskservice.models import JobState
 from controllers.kafka_controller import KafkaController
 from test_common import config
